requirements.txt
transformers==4.44.2
requests~=2.32.3
gspread~=6.1.2
oauth2client~=4.1.3
google-api-core>=1.7,<2dev,!=2.0.*,!=2.1.*
pip~=24.2
telegram~=0.0.1
python-dotenv == 0.20.0
tornado~=6.4.1

tornado~=6.4.1

aiogram~=3.13.0

aiohttp~=3.10.5

torch~=2.4.1
